Fetching live orders during calculation
Need to implement  a wrapper (for caching and seearching) for the Market History crest endpoint.
  * This will provide the adjusted cost which is used for installation fee's

Need to set the adjusted cost in the ActivityMaterialWithCost objects inside the build calculation service (the partially finished method that was just introduced.

+++ aba-market/market-fetcher/src/main/java/com/aba/market/fetch/MarketOrderFetcher.java
@@ -35,5 +35,5 @@ public interface MarketOrderFetcher {
      * @param quantity
      * @return
      */
-    Double getPriceForQuantity ( long regionId, long systemId, long itemId, int quantity );
+    Double getPriceForQuantity ( long regionId, long systemId, long itemId, long quantity );
 }
